restructuring of code for release
